Added test for preparing computations functionality

+++ src/domain/ray.rs
@@ -13,8 +13,8 @@ impl Ray {
     }
 
     // calculates points at given position
-    pub fn position(&self, distance: f32) -> Point {
-        &self.origin + &(&self.direction * distance as f64)
+    pub fn position(&self, distance: f64) -> Point {
+        &self.origin + &(&self.direction * distance)
     }
 
     // transforms ray (building new instance)
